requirements.txt
apache-airflow==2.10.1
boto3
psycopg2-binary
requests
pandas
dbt-core
dbt-redshift
streamlit
prefect
python-dotenv
